requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
astroid==3.1.0
attrs==23.2.0
dill==0.3.8
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
isort==5.13.2
Markdown==3.6
mccabe==0.7.0
multidict==6.0.5
platformdirs==4.2.0
pycodestyle==2.11.1
pylint==3.1.0
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
yarl==1.9.4
